+++ pages/api/get-cms.js
@@ -7,9 +7,8 @@ const mmd = require('micromarkdown');
 const inst = new AirtablePlus({
   baseId: process.env.AIRTABLE_BASE_ID,
   apiKey: process.env.AIRTABLE_API_KEY,
-  tableName: 'CMS'
+  tableName: 'CMS',
 });
-
 
 
 // This takes the data from our CMS airtable
@@ -17,8 +16,10 @@ const inst = new AirtablePlus({
 export default async (req, res) => {
   const records = await inst.read({ maxRecords: 30 });
   res.send({
-    records: records.map(record => {
-      const { key, title, body_en, body_es, picture, secondary_en, secondary_es, href } = record.fields;
+    records: records.map((record) => {
+      const {
+        key, title, body_en, body_es, picture, secondary_en, secondary_es, href,
+      } = record.fields;
       return {
         key,
         title,
@@ -27,17 +28,15 @@ export default async (req, res) => {
         picture: picture || [],
         secondary_en,
         secondary_es,
-        href
+        href,
       };
-    })
+    }),
   });
-
-}
+};
 
 // This adds a <br /> tag where a \n is
 function nl2br(str, replaceMode, isXhtml) {
-
-  var breakTag = (isXhtml) ? '<br />' : '<br>';
-  var replaceStr = (replaceMode) ? '$1' + breakTag : '$1' + breakTag + '$2';
-  return (str + '').replace(/([^>\r\n]?)(\r\n|\n\r|\r|\n)/g, replaceStr);
+  const breakTag = (isXhtml) ? '<br />' : '<br>';
+  const replaceStr = (replaceMode) ? `$1${breakTag}` : `$1${breakTag}$2`;
+  return (`${str}`).replace(/([^>\r\n]?)(\r\n|\n\r|\r|\n)/g, replaceStr);
 }
